+++ cmd/server-manager/typescript/src/Font.ts
@@ -47,6 +47,7 @@ import {faPuzzlePiece} from "@fortawesome/free-solid-svg-icons/faPuzzlePiece";
 import {faBalanceScale} from "@fortawesome/free-solid-svg-icons/faBalanceScale";
 import {faUsers} from "@fortawesome/free-solid-svg-icons/faUsers";
 import {faShieldAlt} from "@fortawesome/free-solid-svg-icons/faShieldAlt"
+import {faVideo} from "@fortawesome/free-solid-svg-icons/faVideo";
 
 library.add(
     faUser,
@@ -95,6 +96,7 @@ library.add(
     faBalanceScale,
     faUsers,
     faShieldAlt,
+    faVideo,
 );
 
 dom.watch();